UX bug: back_up_to is hard to use
E.g. I naturally tried to do something like

    back_up_to foo-bar --exclude '*/Data.fs*' /foo/bar/baz /foo/bar/qux

and it's wrong!  the 2nd arg needs to be a pathname (with leading slash), other args that are pathnames need to drop the leading slash:

    back_up_to foo-bar /foo/bar/baz foo/bar/qux --exclude '*/Data.fs*'

and this is annoying and inaccurate (says it's backing up /foo/bar/baz in verbose mode, when it's actually backing up more).